refactor: style 파일 이동 및 type safety하게 수정 #3
