--- src/assets/theme.ts
@@ -1,20 +1,0 @@
-import "@emotion/react";
-import { colors } from "./colors";
-import { Theme } from "@emotion/react";
-
-const heading1 = {
-  // heading 1 예시입니다.
-  fontWeight: "bold",
-  fontSize: "28px",
-  lineHeight: "150%",
-  letterSpacing: "0",
-};
-
-export const theme: Theme = {
-  colors: {
-    primary: colors.blue,
-    secondary: colors.yellow,
-    warning: colors.red,
-  },
-  heading1,
-};
